Remove ParentNode::match_local again
It was not possible to make it work on Windows. The solution we have here
is also better in the way that it uses already existing functionality.

Some more adjustments were made in order to make the test pass on Windows.
Especially the move contructor on TableView was not working, but this was
not used on Linux as we have copy elision here.

And then some windows header defines "small" as something.

+++ src/realm/query.cpp
@@ -911,7 +911,7 @@ R Query::aggregate(R (ColType::*aggregateMethod)(size_t start, size_t end, size_
             for (size_t t = 0; t < m_view->size(); t++) {
                 ConstObj obj = m_view->get(t);
                 if (t >= start && t < end && eval_object(obj)) {
-                    st.template match<action, false>(obj.get_key().value, 0, obj.get<T>(column_ndx));
+                    st.template match<action, false>(size_t(obj.get_key().value), 0, obj.get<T>(column_ndx));
                     if (st.m_match_count >= limit) {
                         break;
                     }
