+++ src/realm/table_view.cpp
@@ -456,7 +456,7 @@ void TableViewBase::to_json(std::ostream& out) const
         if (key != realm::null_key) {
             if (r > 0)
                 out << ",";
-            m_table->to_json_row(key.value, out); // FIXME
+            m_table->to_json_row(size_t(key.value), out); // FIXME
         }
     }
 
@@ -481,7 +481,7 @@ void TableViewBase::to_string(std::ostream& out, size_t limit) const
     while (count) {
         Key key = get_key(count);
         if (key != realm::null_key) {
-            m_table->to_string_row(key.value, out, widths); // FIXME
+            m_table->to_string_row(size_t(key.value), out, widths); // FIXME
             --count;
         }
         ++i;
@@ -506,7 +506,7 @@ void TableViewBase::row_to_string(size_t row_ndx, std::ostream& out) const
     // Print row contents
     Key key = get_key(row_ndx);
     REALM_ASSERT(key != realm::null_key);
-    m_table->to_string_row(key.value, out, widths); // FIXME
+    m_table->to_string_row(size_t(key.value), out, widths); // FIXME
 }
